add parser

+++ libft/get_next_line/libft.h
@@ -31,7 +31,7 @@ int				ft_isascii(int c);
 int				ft_isdigit(int c);
 int				ft_isprint(int c);
 
-unsigned long	ft_strlen(char *str);
+int				ft_strlen(char *str);
 
 void			*ft_memset(void *s, int c, size_t n);
 void			ft_bzero(void *s, size_t n);
